requirements.txt
PyMuPDF==1.25.1
sentence-transformers==3.3.1
numpy>=1.24.0
groq==0.13.0
python-dotenv==1.0.1
